Add keepalive workflow
GitHub Actions cron jobs only run if the repo had a commit in the last
60 days.  If we had been using a gh-pages branch for deployment, this
would not be an issue.  But using actions to deploy to GitHub Pages
without using a branch means the code does not change all that often.

This keepalive workflow runs on cron, but pushes an empty commit to
the repo every month on the first of the month.  This should keep all
cron jobs running.