Fix commit log loading for unusual tags
Not every tag relates to its branch in the usual way.  (Starting a few days ago.)  So we should fall back to pulling the commit log from the tag ref instead of always using the branch.

We should still try to use the branch commit log, though, because it will be cached from the GitHub API.  If we always use the tag ref, it is both very slow and risks triggering a rate limit on the API.